Discharge, precipitation and temperature in the Chirchik river basin, Uzbekistan.

A dataset containing the 10-day (decadal) or monthly data for discharge, precipitation and temperature in the Chirchiq river basin, Uzbekistan.

ChirchikRiverBasin

Format

A data frame with 29892 rows and 14 variables:

date

Date: Date in decadal or monthly time steps.

data

num: Value of variable.

norm

num: Multi-year average of values.

units

chr: Unit of values. Can be m^3/s (m3s) for discharge, mm (mm) for precipitation or degrees Celsius (degC) for temperature.

type

Factor: Parameter measured. Can be "Q" for discharge, "P" for precipitation or "T" for temperature.

code

chr: Unique station identifier issued by Uzbek HydroMet.

station

chr: Station name.

river

chr: Name of river the station is related to.

basin

chr: The name of the basin the station is part of.

resolution

chr: Temporal resolution of data. Can be dec for data in 10-day intervals (decadal) or mon for monthly intervals.

lon_UTM42

num: Longitude of station in UTM42.

lat_UTM42

num: Latitude of station in UTM42.

altitude_masl

num: Altitude of station in meter above mean sea level.

basinSize_sqkm

num: Size of the basin in square kilometers (km2).

Source

Uzbek HydroMet 2021, Andrey Yakovlev