Assesses the quality of decadal or monthly forecasts as done in Central Asian Hydrometeorological Agencies.

This function computes a per period (decade or month) forecast quality assessment for each forecast for the test set provided. Ideally, the user provides an out-of-sample testset

assess_fc_qual(df, plot)

Arguments

df

tibble with date, obs, pred and per column where the per column is either the decade or the month specifier (normally given as factor).

plot

TRUE/FALSE if plot should be generated

Value

A list with a tibble of per period forecasts that are acceptable expressed as percentages of the total test set length