Creates a vector of decadal (10 days) dates

This function creates a decadal 10-days dates vector that allows date tagging for decadal (10-days) time series. This type of timeseries is usually used for hydro-meteorological data in the former Soviet Republics. The intra-months decades can be configured as dates at the beginning or end of the decade.

decadeMaker(s, e, type)

Arguments

s

starting date in YYYY-mm-dd format

e

end date in YYYY-mm-dd format

type

'start' creates starting decade dates, 'end' creates ending decade dates.

Value

A sequence of decadal dates

See also