Generate hourly precipitation and temperature time series from daily data

The function uses different gap filling methods to a) compute hourly precipitation, and b) compute hourly temperatures for all stations/elevation bands. Daily precipitation values are uniformly distributed over the day and temerpatures are calculated to fit a typical diurnal station cycle between daily Tmin and Tmax (output of stochastic weather generator). The diurnal cycle can be calculated by the function diurnalTempCycle_stations().

generateHourlyFromDaily_PT(PT_sim, era5_data, param)

Arguments

PT_sim
Climate scenario dependent daily precipitation and temperature station values

era5_data
Full ERA5 data frame

param
Stochastic weather simulation parameters

Value

List with climate scenario dependent hourly P and T data as direct input for RS MINERVE.