Extract subbasin-level and elevation band-specific hourly time series from ERA5 raster bricks

Function extracts precipitation (tp) or temperature (t2m) data from raster bricks and prepares a dataframe for later import in RSMinerve (use write.table() function to export as csv). If tp and t2m have to be exported, the function has to be called twice and the resulting tibble columns added.

generate_ERA5_Subbasin_CSV(
  dir_ERA5_hourly,
  catchmentName,
  dataType,
  elBands_shp,
  startY,
  endY
)

Arguments

dir_ERA5_hourly

Path to original stored DEM file (generate_ERA5_Subbasin_CSV assumes the following folder structure: <dir_ERA5_hourly>/<tp or t2m>/<catchmentName>/*.nc)

catchmentName

Name of catchment for which data should be extracted (depends on proper local file storage)

dataType

Currently, only 'tp' or 't2m' as valid options

elBands_shp

Subbasin-level shapefile with elevation bands

startY

Starting year for which data should be made available (assuming data 'is' available from the start of that year)

endY

Ending year from which data should be extracted (assuming data 'is' actually available until the end of that year)

Value

Dataframe tibble with temperature in deg. C. or precipitation in mm/h

See also