Calculate glacier mass balance of all glaciers in a basin over time.

Small glaciers below area_threshold in km2 are treated as one unit whereas large glaciers above area_threshold are treated in elevation bands.

glacierBalance(melt_a_eb, rgi_elbands, area_threshold = 1)

Arguments

melt_a_eb
Annual glacier melt in mm/a for each glacier/elevation band. The aggregation period is from October 1 of the previous year to September 30 of the current year. The object melt_a_eb must contain the columns Hyear for the hydrological year, ID for the glacier ID in the RGI format and Melt with the glacier melt in mm/a. The column ID is generated by paste0(RGIId, "_", elvtn_b) in the input rgi_elbands.

rgi_elbands
A shape file based on the Randolph Glacier Inventory but with the larger glaciers subdivided into elevation bands. The object rgi_elbands must contain the columns RGIId, A_km2 with the area of the glaciers/elevation bands in km2, thickness_m with the thickness of the glaciers/elevation bands in m. We further need a column called Area_tot_glacier_km2 with the area of the entire glacier within the catchment and a column ID which is consistent with the ID column in object melt_a_eb.

area_threshold
Threshold area separating small and large glaciers in km2. The default is 1 km2. The threshold value must be consistent with the one used to generate the rgi_elvands object.

Value

A tibble containing annual glacier melt, imbalance ablation, glacier volume, glacier area.

Details

The input object melt_a_eb can be generated by aggregating the daily glacier melt obtained with the function glacierMelt_TI() to the hydrological year. The vignette glaciers-02 demonstrates a use case for the function.

See also

Examples