Calculates glacier melt and the root mean squared error between observed glacier melt and simulated glacier melt.

This function calls ```glacierMelt_TI``` with ```parameters``` and ```temperature``` and calculates the root mean squared error RMSE between the simulated and observed ```observed``` glacier melt. The optional ```index``` serves to limit the results to one single glacier. \ This function can be used to calibrate the parameters of the temperature index model. As the recommended optimization algorithm GA is a maximiser, the function returns the negative RMSE.

glacierDischargeRMSE(
  parameters,
  temperature,
  observed,
  index = 1:(length(unique(observed$RGIID)))
)

Arguments

parameters

The two parameters of the function ```glacierMelt_TI```. The first parameter is the melt factor ```MF```, the second parameter is the threshold temperature ```threshold_temperature``` above which melt sets on. Example: c(4, 0).

temperature

A tibble with the time steps in rows and temperature values for individual hydrological response units in columns. The names of the hydrological response units is typically of the form RGIId_layer, e.g. RGI60-13.10007_1. The RGIId and the underbar are required as the algorithm splits the name at the underbar and mergest the observed glacier melt by RGIId.
The dates column of the temperature time series can be in the format lubridate::date for sub-annual glacier melt simulations or an integer indicating the (hydrological) year for which glacier melt is calculated. The date column must be named ```date``` for sub-annual time steps or ```year``` for annual time steps.

observed

A tibble with the glacier melt data by Miles et al. ```observed``` must contain the columns RGIID, totAbl and Area_m2. If a column ```year``` is present, the algorithm assumes yearly data, else it assumes that ```observed``` represents average data over the entire period of the temperature time series.

index

(optional) number to indicate for which individual HRU the RMSE should be returned. Defaults to all.

Value

-1*RMSE between observed glacier melt and glacier melt simulated with the temperature index model of indexed hydrological response units.

See also