Calculates glacier melt and the root mean squared error between observed glacier melt and simulated glacier melt.

This function calls ```glacierMelt_TI``` with ```parameters``` and ```temperature``` and calculates the root mean squared error RMSE between the simulated and observed ```hugonnet``` glacier melt. The optional ```index``` serves to limit the results to one single glacier. \ This function can be used to calibrate the parameters of the temperature index model. TODO write vignette demonstrating use.

glacierRMSE(
  parameters,
  temperature,
  hugonnet,
  index = 1:(dim(temperature)[2] - 1)
)

Arguments

parameters

The two parameters of the function ```glacierMelt_TI```. The first parameter is the melt factor ```MF```, the second parameter is the threshold temperature ```threshold_temperature``` above which melt sets on. Example: c(4, 0).

temperature

A tibble with the time steps in rows and temperature values for individual hydrological response units in columns. The names of the hydrological response units is typically of the form RGIId_layer, e.g. RGI60-13.10007_1. The RGIId and the underbar are required as the algorithm splits the name at the underbar and mergest the observed glacier melt by RGIId.

hugonnet

A tibble with the yearly glacier melt data by Hugonnet et al. ```hugonnet``` must contain the columns rgiid, start and dmdtda. Further columns are ignored.

index

(optional) number to indicate for which individual HRU the RMSE should be returned. Defaults to all.

Value

RMSE between observed glacier melt and glacier melt simulated with the temperature index model of indexed hydrological response units.

See also