Import iEasyHydro Data

Loads csv files that were exported from iEasyHydro. The function automatically computes the long-term norm of the data provided and returns a time aware tibble with date, data, data norm columns and code columns, etc. This function is an adapted version of the riversCentralAsis::loadTabularData() function.

import_iEasyHydro_data(
  fPath,
  fName,
  code,
  stationName,
  rName,
  rBasin,
  dataType,
  units,
  x_UTM42N,
  y_UTM42N,
  masl,
  basin_size
)

Arguments

fPath

Path to the input file

fName

File name

code

Hydrometeorological station code

stationName

Hydrometeorological station name/location

rName

Name of river

rBasin

Name of basin

dataType

Type of data, either `Q` (discharge data), `T` (temperature data) or `P` (precipitation data)

units

Data units

x_UTM42N

x coordinates in UTM42N

y_UTM42N

y coordinates in UTM42N

masl

station elevation

basin_size

basin size in km^2

Value

Time-aware tibble with relevant data

See also