Reads climate and discharge data from a character string into a tibble.

The data is in a string in the same format as the RSMinerve forcing csv that can also be imported to RS Minerve to form a database. This function is useful for reformatting data in character format that is to be written to RS Minerve input files.

readForcingSTR(string, tz = "UTC")

Arguments

string

Character string with input data for RS Minerve.

tz

Time zone character to be passed to lubridate::as_datetime. Defaults to "UTC"

Value

Returns a tibble of the same format as data with data in hourly (climate) to decadal or monthly (discharge) time steps. Includes all attributes of the csv file.

See also