Estimates glacier surface area and glacier volume.

AizensGlacierFunctions implements the empirical relationships between glacier length, surface area and volume found by Aizen et al. in the Tien Shan mountains. They can be used to estimate regional glacier volumes.

AizensGlacierFunctions(glaciers)

Arguments

glaciers

A data.frame, tibble, or sf object. Must contain an attribute ```Lmax``` in meters (as in the Randolph Glacier Inventory (RGI 6.0)) or ```Length [km]``` in kilometers or ```L(t) [km]``` (as in the output from Oerlemans glacier length function).

Value

Returns the input object with Area_Aizen_km2 and Volume_Aizen_km3 added. Will return NULL if the length column in the input object is not found.

References

Aizen, Aizen & Kuzmichonok (2007) Glaciers and hydrological changes in the Tien Shan: simulation and prediction. Environmental Research Letters. DOI: 10.1088/1748-9326/2/4/045019.

See also

Examples

glacier <- tibble::tibble(Year = c(2000:2010),
                          "Length [km]" = c(5.7, 5.6, 5.5, 5.5, 5.4, 5.3,
                                            5.3, 5.2, 5.1, 5, 4.8))
AizensGlacierFunctions(glacier)
#> # A tibble: 11 x 4
#>     Year `Length [km]` Area_Aizen_km2 Volume_Aizen_km3
#>    <int>         <dbl>          <dbl>            <dbl>
#>  1  2000           5.7           8.90            0.606
#>  2  2001           5.6           8.62            0.580
#>  3  2002           5.5           8.35            0.554
#>  4  2003           5.5           8.35            0.554
#>  5  2004           5.4           8.08            0.529
#>  6  2005           5.3           7.82            0.505
#>  7  2006           5.3           7.82            0.505
#>  8  2007           5.2           7.55            0.482
#>  9  2008           5.1           7.30            0.460
#> 10  2009           5             7.04            0.438
#> 11  2010           4.8           6.55            0.396